New automatic pump code & library
Includes library with all relevant peristaltic pump functions (with room to expand for mixer pump and linear rail functions). AutoStep_0.3 is a basic layout for how the program will run; can be expanded for more pumps by adding more to the array; fits current method for ordering but easy to change.